% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliseResultsTable.R
\name{addMissingMin}
\alias{addMissingMin}
\title{Some sites with laps under one minute don't include any minute prefix}
\usage{
addMissingMin(x)
}
\description{
Some sites with laps under one minute don't include any minute prefix
}
