% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliseResultsTable.R
\name{extractSex}
\alias{extractSex}
\title{Add Sex Column to Results Table, extracting from a suggested field}
\usage{
extractSex(resultsTable, suggestedColumnName)
}
\description{
Takes a table of results, and a suggested column to extract the sex from.
}
