% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliseResultsTable.R
\name{highlightNames}
\alias{highlightNames}
\title{Add highlight and highlightSex Column to Results Table}
\usage{
highlightNames(resultsTable, suggestedNameColumn, namesToHighlight)
}
\description{
Takes a table of results, a suggested column to match and a vector to match against.
}
