% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliseResultsTable.R
\name{normaliseResultsTable}
\alias{normaliseResultsTable}
\title{Normalise Results Table}
\usage{
normaliseResultsTable(rawResultsTable, guessedStageNames = c())
}
\description{
Takes a table of results, one column per lap/stage, one row per rider.
renames numbered lap and named stage columns to numbered stages
}
