% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliseResultsTable.R
\name{stageTimesToSeconds}
\alias{stageTimesToSeconds}
\title{Tidy up the stage time strings, and convert to seconds '_lub' suffix}
\usage{
stageTimesToSeconds(normResultsTable)
}
\description{
Tidy up the stage time strings, and convert to seconds '_lub' suffix
}
