% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliseResultsTable.R
\name{stripRank}
\alias{stripRank}
\title{RootsAndRain shows the rank within each stage in the stage time - keep only the time info}
\usage{
stripRank(x)
}
\description{
RootsAndRain shows the rank within each stage in the stage time - keep only the time info
}
